Code for specific machine
Update the kerf width for gcode generation.  Modified gcodegenerator
file by removing all the zaxis commands. Also ajusted velocity for
circles.

Removed a bit of code that returned the torch to the beginning of the
block. IOW, the code use to let the machine cut back on the intro line,
which is not nice.'